Fix IDEA-224511: (NPE) add null check for classData object
When running unit tests with coverage, sometimes the method
ProjectData.touchLine() method is called with a null object.

When looking at the NPE stacktrace, it always is associated
with a class from the a maven sources jar (*-sources.jar).

To reproduce: add all the *-sources.jar and *-sources.jar.sha1
for all external maven dependencies to your maven repository
and run a test-with-coverage.

- Add a null check for classData object inside touchLine()